+++ guppylang/src/guppylang/decorator.py
@@ -78,9 +78,12 @@ AnyRawFunctionDef = (
     OverloadedFunctionDef,
 )
 
-__all__ = ("GuppyKwargs", "custom_guppy_decorator", "guppy", "link_name", "metadata")
-
-LINK_NAME_KEY = "link_name"
+__all__ = (
+    "GuppyKwargs",
+    "custom_guppy_decorator",
+    "guppy",
+    "metadata",
+)
 
 
 class GuppyKwargs(TypedDict, total=False):
@@ -694,25 +697,6 @@ def metadata(key: str, value: Any) -> Any:
     return decorator
 
 
-def link_name(name: str) -> Any:
-    """Decorator to attach a link name to a Guppy definition. It must be placed below
-    the @guppy decorator.
-
-    .. code-block:: python
-
-        from guppylang import guppy
-        from guppylang.decorator import link_name
-
-        @guppy.declare
-        @link_name("my_link_name")
-        def main() -> None:
-            pass
-
-        main.compile()
-    """
-    return metadata(LINK_NAME_KEY, name)
-
-
 def _parse_expr_string(ty_str: str, parse_err: str, sources: SourceMap) -> ast.expr:
     """Helper function to parse expressions that are provided as strings.
 
@@ -865,10 +849,10 @@ def _parse_kwargs(kwargs: GuppyKwargs) -> ParsedGuppyKwargs:
     if "max_qubits" in kwargs:
         metadata.set_max_qubits(kwargs.pop("max_qubits"))
 
-    if LINK_NAME_KEY in kwargs:
+    if "link_name" in kwargs:
         raise TypeError(
-            "`link_name` keyword argument is not allowed in @guppy decorator, "
-            "use @link_name decorator instead"
+            "`link_name` keyword argument has been removed from the `@guppy` decorator,"
+            " use the `@link_name` in `guppylang.library` decorator instead."
         )
 
     if remaining := next(iter(kwargs), None):
@@ -882,10 +866,11 @@ def _parse_kwargs(kwargs: GuppyKwargs) -> ParsedGuppyKwargs:
 
 
 def _get_link_name(f: Callable[..., Any]) -> str | None:
-    """Returns the link name for a Guppy function, if it has one."""
+    from guppylang.library import LINK_NAME_METADATA_KEY
+
     custom_metadata = getattr(f, "__guppy_metadata__", {})
     assert isinstance(custom_metadata, dict)
-    return custom_metadata.get(LINK_NAME_KEY, None)
+    return custom_metadata.get(LINK_NAME_METADATA_KEY, None)
 
 
 @hide_trace
@@ -896,8 +881,6 @@ def _add_generic_metadata(f: Callable[..., Any], metadata: FunctionMetadata) -> 
     custom_metadata = getattr(f, "__guppy_metadata__", {})
     assert isinstance(custom_metadata, dict)
     for key, value in custom_metadata.items():
-        if key == LINK_NAME_KEY:
-            continue  # link_name is handled separately
         metadata.set_generic_metadata(key, value)
