+++ tests/integration/library/test_struct.py
@@ -1,5 +1,6 @@
 from guppylang import guppy
-from guppylang.decorator import link_name
+from guppylang.library import link_name
+
 from guppylang.library import GuppyLibrary
 from guppylang.std.platform import output
